[Build] Refactor project structure and refactor release tools
