+++ .github/workflows/check-java-bridge-licensing.yml
@@ -40,14 +40,14 @@ jobs:
       - name: Build
         run: |
           set -o pipefail
-          cd pypaimon/py4j/paimon-python-java-bridge
+          cd paimon-python-java-bridge
           mvn clean deploy ${{ env.MVN_COMMON_OPTIONS }} -DskipTests \
             -DaltDeploymentRepository=validation_repository::default::file:${{ env.MVN_VALIDATION_DIR }} \
             | tee ${{ env.MVN_BUILD_OUTPUT_FILE }}
 
       - name: Check licensing
         run: |
-          cd pypaimon/py4j/paimon-python-java-bridge
+          cd paimon-python-java-bridge
           mvn ${{ env.MVN_COMMON_OPTIONS }} exec:java@check-licensing -N \
             -Dexec.args="${{ env.MVN_BUILD_OUTPUT_FILE }} $(pwd) ${{ env.MVN_VALIDATION_DIR }}" \
             -Dlog4j.configurationFile=file://$(pwd)/tools/ci/log4j.properties
