--- pypaimon/py4j/paimon-python-java-bridge/src/main/resources/META-INF/NOTICE
@@ -1,21 +1,0 @@
-paimon-python-java-bridge
-Copyright 2024 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-This project bundles the following dependencies under the under the MIT license
-- org.slf4j:slf4j-api:1.7.32
-
-This project bundles the following dependencies under the Apache Software License 2.0 (http://www.apache.org/licenses/LICENSE-2.0.txt)
-- org.apache.logging.log4j:log4j-1.2-api:2.17.1
-- org.apache.arrow:arrow-vector:14.0.0
-- org.apache.arrow:arrow-format:14.0.0
-- org.apache.arrow:arrow-memory-core:14.0.0
-- org.apache.arrow:arrow-memory-unsafe:14.0.0
-- org.apache.arrow:arrow-c-data:14.0.0
-- com.google.flatbuffers:flatbuffers-java:1.12.0
-
-This project bundles the following dependencies under the 2-Clause BSD License
-- net.sf.py4j:py4j:0.10.9.7
-
